pyproject.toml
[project]
name = "ecatalog-cli"
version = "0.1.0"
description = "CLI tool for eCatalog API workflows and operations"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.1",
    "xlrd>=2.0.2",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
ecatalog = "cli:main"

[tool.uv]
dev-dependencies = []
